Functional Python backend
