+++ utils/polus-bfio-util-py/VERSION
@@ -1,1 +1,1 @@
-2.4.4
+2.0.0
