--- utils/polus-bfio-util-py/bfiop/__init__.py
@@ -1,3 +1,0 @@
-from __future__ import absolute_import, unicode_literals
-
-from .bfio import BioReader,BioWriter
